Preferences menu showing on preferences
