--- Brooklyn/Contexts/Root/PreferencesTabController.swift
@@ -1,19 +1,0 @@
-//
-//  PreferencesTabController.swift
-//  BrooklynApplication
-//
-//  Created by Pedro Carrasco on 19/02/2019.
-//  Copyright © 2019 Pedro Carrasco. All rights reserved.
-//
-
-import Cocoa
-
-final class PreferencesTabController: NSTabViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        addChild(AnimationsViewController(nibName: AnimationsViewController.identifier, bundle: nil))
-
-        tabViewItems.first?.label = "Animations"
-    }
-}
